Fix #17 and fix #18
Add tests to the shared test base for these and make sure ordered
container runs them.
